[ConfigAPI] Refactor Events and improve ConfigAPI
https://docs.python.org/2/library/threading.html#event-objects

With the current implementation of the event classes it is not possible
to see the result of a channel switch. The new event classes have a base
class that holds a threading.Event. Subclassing was unfortunately not
possible. The threading.Event allows waiting for an event to be set().
The base class holds also attributes for result (True/ False) and a
message. After the fronthaulemulator triggered the channel switch we can
now forward the result from the ConfigAPI.