[iperf] Implement iperf clients and servers
Introduce Node/Host classes that allow spawning iperf servers and
clients to generate traffic in the terranet. Iperf servers and clients
are started in the terranet on startup. In the example network the
servers are behind the gateway and each ClientNode gets two iperf
clients assigned, that emulate users behind the client nodes.