Build komondor config from topo (#5)
* [Wifi] Implement wifi channel to komondor mapping

https://en.wikipedia.org/wiki/List_of_WLAN_channels

Some of the channel combinations we have simulated are not allowed.
Implement a mapping between wifi channels and komondor settings.

* [wifi] Adjust config API to the channel module

The previous commit implemented a wifi channel abstraction. Adjust the
config API to take a channel number as parameter.

* [komondor] Implement komondor config classes

Instead of using vanilla dicts implement komondor config and config
section classes.

* [komondor] Move komondor class to wifi package

Move komondor module to wifi package. Add method to call parallel
komondor simulation from the fronthaul emulator.

* [Refactor] Rename topo methods

Rename the long iper client and server methods to something shorter.

* [link] Rename link classes

* [fronthaul emulator] Refactor to build from topo

Refactor the fronthaul emulator to build the komondor configuration of
the wifi network. This centralizes the komondor simulation at the
fronthaul emulator component. Makes it easier to write new terranet
topologies while adding complexity to the fronthaulemulator.

* [refactor] Adjust net and node module

Adjust node and net modules to the refactoring.

* [examples] Replace the old examples

The advantage of the refactoring introduced by the former commits is
that we can now implement the whole terranet in a single topo file. This
follows the design concept of mininet and makes it easier to simulate
new terranet topologies.

* [scripts] Remove scripts

The scripts for the multistep simulation are not needed since the
fronthaul emulator parses the komondor configuration from the topo file.

* [examples] Fix examples

Fix wrong superclass call in the examples.

* [examples] Remove unnecessary build call

Start builds the topo.

* [refactor] Divide huge build method

The build komondor method is huge and complicated. Factor out some
methods to simplify code.