[controller] Add first controller version
Attention - This version of the controller doesn't work yet. Problems
detected so far:

1. The table-miss action for flodding in the last table of the pipeline
   doesn't work (code: OFPBMC_BAD_PREREQ(9)).
2. The MAC learning doesn't work yet. Packets with same src MAC hit the
   learning table several times.

Add a controller that is able to detect customer flows that are
identified by a certain prefix. The matches of the customer IPv6 src
addresses is only to identify connected customer CPEs and monitor their
traffic.

Replace the OVS default learning by a OpenFlow based pipelined
solution.