[Proxy] Implement a namespace proxy
Implement a proxy for ConfigurableWifiAccessPoint. This proxy allows to
send HTTP requests to the ConfigAPIs of the AccessPoint. The
AccessPoints are running in network namespaces and can not be addressed
from outside the network. Unfortunately, our RYU controllers are running
on the host machine. The namespace proxy forwards requests from the host
machines into the network namespaces to provide access to the ConfigAPIs
of the AccessPoints.