[Vagrant] Add aws provider and packer provisioning
Move the vagrant provisioning to packer. Create a vagrant folder at the
top level and with hierarchical Makefile structure to build the vagrant
boxes. This is a lot of work in the first place, but eases up the build
pipeline for the aws provider.

Add a build pipeline for the vagrant aws provider to allow simulations
on big AWS machines.